+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Changelog
 * **Enhancement**: Transform and collect contribution data in a single pass during IR.
 * **Fix**: Ensure contributed binding containers' included containers are available in root dependency graphs.
 * **Fix**: Make `@Includes` parameter keys available to extensions.
+* **Fix**: Report diagnostic errors if accessors or injectors have conflicting qualifiers in overridden functions.
+* **Fix**: Report diagnostic errors if an injector function does not return `Unit`.
 
 0.6.3
 -----